Add specific rate limiting for licensify
The TurnitinBot seems to be making a few to many requests, the rate is
overall around 3 to 4 per second, which is seemingly too many for
licensify, but not enough to be caught by the rate limiting on the
cache machines.

Also, because there are currently 9 cache machines, it's hard to rate
limit there, as the limits have to be set per individual machine.

Therefore, do some rate limiting closer too the app, which should be
more effective as there are currently only two licensify-frontend
machines. Set a low limit of 1r/s so that the limit across the two
machines is around 2r/s, but set a large burst queue length of 40 so
that it takes quite a few requests in a short period of time for this
to start taking effect.