Upgrade Mongo On CI Agents 1 - 3
Upgrade Mongo to 2.6 from 2.4.9 on ci_agent machines 1 - 3. This
is a necessary step towards migrating the remaining publishing apps
currently using Mongo databases from Carrenza to AWS. We will be updating the apps
to use DocumentDB which is incompatible with versions of Mongo below 3.6. We
are upgrading the CI Agent Machines to bring the versions inline with the versions
used in all other environments. For more information please see this trello card:
https://trello.com/c/YXXu4s5c/3422-fix-mongo-documentdb-errors