Fall back to rubygems if gem is not in gemstash
... or if our gemstash goes down for some reason.

Pretty sure we're on > 1.12.x everywhere now, and if not this will force
us to fix it.

Note that the name of this config setting is slightly different to the
one in the comment - `/__FALLBACK...` instead of `.FALLBACK`. This is
what I get in my `~/.bundle/config` if I run `bundle config mirror.https://rubygems.org.fallback_timeout true`,
which is what the docs say I should do to set this flag.