Update ruby 1.9.3's bundler_version to 1.17.3
This is the latest version of bundler that works with ruby < 2.

In practice, I don't think this version of bundler is actually used
anywhere, because ruby 1.9.3 should only be used for running puppet,
which doesn't run `bundle`. It feels safer to upgrade it though, as
we're now using some features of newer versions of bunlder (falling back
to rubygems if gemstash doesn't have a gem).