Remove backdrop from our deploy infrastructure
While this app is still provisioned on our infrastructure, we have
been instructed to only deploy it using the legacy alphagov-deployment
repo. This removes 'backdrop-read' and 'backdrop-write' from the list
in the Deploy_App job, since we must not deploy it this way. Note this
is a Python repo, so also has no 'rake' tasks.