Proxy draft assets to their respective draft app
Previously all of the assets used in the draft stacks for frontend apps
were live ones and this relied on both the draft and live assets
generating the exact same signatures on assets.

With the switch from assets hostname to the www one any assets with
absolute URLs (including hosts) are producing slightly different assets
on draft and live stacks and thus have different signatures. This has
meant that assets have been 404ing.

I've moved this configuration to be at the class level to avoid
situations of forgetting to set this in draft stack and it working
nearly all the time based off the common configuration.