Configure whitehall to allow assets from www hostname
This sets up assets/whitehall as a path where Whitehall can write it's
assets. This marks a bold new world outside of the confines of the
/government prefix (Whitehall assets are currently served from
/government/assets). This path is then set-up to proxy in live and draft
router. The draft proxy is to draft-whitehall-frontend which is just an
alias to whitehall-frontend, however this seems to be the more
conventional hostname used on the draft stack.