Disable Faulty Carrenza Production to AWS Staging Sync of Manuals Publisher Database
Due to an incompatibility between DocumentDB and Mongo 2.6 the govuk_env_sync
job is failing to restore the Manuals Publisher database. The sync job errors
with "error running create command: Field 'create' is currently not supported"
We are temporarily disabling the sync for Manuals Publisher to stop the AWS
Staging Manuals Publisher database being wiped every night by the failed sync,
but we will ensure this is fixed before the migration.