Configure feedback to allow assets from www hostname
This was missed in https://github.com/alphagov/govuk-puppet/pull/10385

This sets up assets/feedback as a path assets can be wrote to
for the feedback app. This then sets up that path in router to proxy
requests for that app. It doesn't set up a proxy in draft-router as
feedback only runs in the live stack.