Add redirect for Whitehall assets on assets host
Whitehall does a rather strange where it embeds it's asset URLs into
some content items when they use a govspeak asset. This means that,
unlike other asset URLs, some of these need to be maintained
permanently. To handle maintaining these permanently this code adds a
redirect on the asset hostnames so any requests for
files to /government/assets are redirected.

In investigating how much this was needed I only found one case of a
document that needed this [1] in the content store, all the other
assets are configured in a strange way that I deal with in the next commit.
Looking in Athena there was still a reasonably high amount of requests
for assets so I think it's still best to keep it for now as it will
prevent some broken looking pages.

I didn't think there'd be any scope for there to be requests to
draft-assets since Whitehall doesn't use that hostname, so I've only
done this for the live asset hostname.

[1]: https://www.gov.uk/government/publications/national-curriculum-in-england-mathematics-programmes-of-study/national-curriculum-in-england-mathematics-programmes-of-study