Sync and restyle Sidekiq job count
Previously the line showing the count of processed Sidekiq jobs used
a different interval to the other lines on the graph, which lead to
confusion since the points on the graph were being aggregated over
different time intervals. This fixes that.

After fixing the time interval, it's hard to distinguish the job count
line from the lines showing the job times. This restyles the line so
it's clear that it represents a different kind of series.