Up totals for daily Email Alert API monitoring
The previous limit of 10,000,000 no longer applies (and I believe this
was actually changed a while ago) instead GOV.UK no longer has it's own
individual limit but Notify does have a limit with SES of 30,000,000.

To reflect this I've upped the limits to reflect 25,000,000 as our
daily limit. Warning threshold is set at 80% of this, 20m, and critical
at 90%, 22.5m.