Disallow Googlebot from crawling licence-finder
We need to do this as Googlebot is spamming licence-finder with millions
of requests a day. We initially tried rate limiting by IP but the variation of
IP addresses google uses renders this approach ineffective.

Google is trying to index every variation of the licence-finder search
results (different query param per variation) which is resulting in a
lot of requests putting strain on the app.

Me and Alan have discussed this and think it's fine to disallow google
from indexing these pages as the main licence-finder start page will
still be indexed. This comes off a suggestion from Jonathan Nichols.