Rate limit per IP address for licencefinder
This PR is a result of trying to limit google's crawler bot activity for
our licencefinder app. We're currently experiencing around 175 r/s
consistently for variations of query string for the path
/licence-finder/sectors and this figure is rising day by day.

This PR will limit single IP addresses from requesting more than twice a
second but with the flexibility of burst requests being allowed so we
don't throttle users who may for example refresh the page multiple times
in a second. This burst limit is set to 40, if the client consistently
requests more than twice a second then they are throttled and a 429 is
returned. More info here -> https://www.nginx.com/blog/rate-limiting-nginx/

This 429 should tell the google bot to crawl less and reduce
the traffic to licencefinder.

This PR has taken inspiration from
https://github.com/alphagov/govuk-puppet/pull/10325/files.