Ensure pingdom monitoring is absent
This monitoring of Pingdom only monitored a few of the Pingdom checks we
have configured, since this isn't comprehensive (and has been this way
for a long time) it indicates we're not getting particularly high value
from it and so have chosen to remove it. This monitoring would have been
better if puppet was also provisioning the checks, but since the
management of them is done through the web interface this puppet code is
just another thing to remember.

This is part of wider work to try bring consistency to our Pingdom
configuration.