start push of email alert api earlier in production
Based on analysis of the govuk_env_sync database backup to S3, email alert api takes more than 1 hour to backup to S3 in production.

09/06: 61 minutes
08/06: 63 minutes
07/06: 64 minutes
06/06: 69 minutes
05/06: 67 minutes
04/09: 68 minutes
03/09: 72 minutes
02/09: 76 minutes

Therefore, we start the backup earlier so that production S3 bucket is updated before the restoration in staging and integration. It should be noted that integration is restored from production because its anonymisation script deletes data more than 1 day old.