Refresh the Email Alert API Metrics dashboard
This refreshes the dashboard to help avoid the need to visit other
dashboards when investigating performance. Notable changes:

- Introduction of a "$Aggregation" variable to consistently summarise
graphs in specified buckets, to avoid visual overload

- Restyling of most graphs to be similar to the Sidekiq dashboard, which
uses thin, lightly faded lines with small points

- Introduction of 'consolidateBy' (even where we don't need it, for
consistency) to ensure we don't lose data when Graphite buckets it.

- Replacing the per-minute Notify dial with a graph, using 30s buckets
to try and expose any peaks in the rolling average.

- Addition of a link to the AWS RDS dashboard, so we have some awareness
of the stats for the Postgres DB we're relying on.

- Removal of noisy 'success' series, in favour of enqueued / processing
stats, and failures (on a separate axis, so they stand out).

- Removal of the bottom row of timing graphs for DeliveryRequestWorker,
since these haven't proven to be useful.

- Addition of 'upper' series for Notify requests and status updates,
as these contribute significantly to 'upper' Delivery Job durations.