Fix misleading 'check_proc_running' alert messages
Some of our alerts were down as 'X not running', as opposed to
'X running'. When the alert happens, the 'Service' message we
see in Icinga/Nagstamon is the 'X running'/'X not running'
message. It's a little confusing to see 'X running' as an alert
when the issue is that it ISN'T running.