Add X-Robots-Tag header when deny_crawlers is true
This is another way of not allowing search engines crawlers from
crawling the site. Crucially, it removes the pages from search results
if they are linked to, even if the domain has a `robots.txt` file which
would ordinarily disallow the crawler.

See
https://support.google.com/webmasters/answer/6062608?hl=en&ref_topic=6061961&visit_id=637278233305826665-2774335487&rd=1
for more information on the different ways of denying access to a
cralwer.