Set an ASSET_HOST env var for Whitehall Frontend
Whitehall frontend has the surprising characteristic that it can be
served from two different hosts: www.gov.uk and
assets.publishing.service.gov.uk. The latter of which is used in the
case of CSV previews (Example [1]), at least until RFC 122 [2] is
implemented. Because of this serving a relative path to assets will fail
when the hostname changes.

To resolve this I've decided to set the asset host absolutely on
Whitehall frontend to the web root. Normally in this situation we'd also
have to set a draft asset host too, but since Whitehall draft is just
the same app proxied this isn't necessary.

[1]: https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/887009/exrates-monthly-0620.csv/preview
[2]: https://github.com/alphagov/govuk-rfcs/blob/master/rfc-122-csv-preview.md