Remove Whitehall's custom ASSET_ROOT env var
Whitehall no longer makes use of this of this environment variable so
this can be removed. It's rather confusing that Whitehall customises
this as I interpret this environment variable to be the asset root for
all of GOV.UK not a particular app.

Anyway once this is cleared out I plan to change Whitehall to use the
regular GOVUK_ASSET_ROOT env var as part of an idea that we can kill off
the need for both GOVUK_ASSET_HOST and GOVUK_ASSET_ROOT.