Activate Continuous Deployment
This adds a conditional step to the Deploy_App job, which will trigger
a downstream deploy for a standard "deploy" build of a specific release
tag of a supported app. Adding all these conditionals means we avoid
triggering invalid builds of the Deploy_App_Downstream job, which would
look alarming if someone views the job logs.

Note that merging a PR always triggers a standard "deploy" build [1],
so we should avoid triggering a downstream deployment for other deploy
tasks; most apps will automatically run migrations as part of "deploy".

[1]: https://github.com/alphagov/govuk-jenkinslib/blob/b5af39d69a14c94bcecb770f7bba40b0ec672af7/vars/govuk.groovy#L134