Reflect Email Alert API worker rename
We've renamed the ProcessContentChangeAndGenerateEmailsWorker to
ProcessContentChangeWorker, this change reflects that in the dashboards.

We could change these to use ProcessContentChange*Worker to capture both
(as I've done on a temporary board) current name and old name. However,
since these dashboards only serve to provide a somewhat short term view
of the apps health it seems a superfluous step for the long term.