Remove redundant subscription content alerts
Previously we created SubscriptionContent records before actual
Email records, with the risk of stopping part way through and
not having emails to send. Recently we inverted this process, so
that SubscriptionContent and Email records are created in the
same transaction [1]. This removes unprocessed SubscriptionContent
alerts, since it's no longer possible for them to fire.

[1]: https://github.com/alphagov/email-alert-api/pull/1284