Add Test class; fix xvdf issue
This copies the integration class over to test.

The motivation is that /dev/xvdf doesn't exist
on the machine, but /dev/nvme1... does.
This follows on from a change made to integration
for the same purpose (https://github.com/alphagov/govuk-puppet/pull/9392)

You may see errors like the following if this change
isn't fixed elsewhere:
Error: Execution of '/sbin/pvcreate /dev/xvdf'...