Limit requests for new email subscriptions
This applies a rate limit to the specific request that causes
subscription confirmation emails to be sent. At the moment, we
are seeing a high number of bot-like requests to this endpoint.
The limit is based on the one used for /contact/* paths, after
realising we can only specify a per-minute granularity:

- 3 requests per minute (180 per hour)
- burst of 0, since we want to have a hard limit

Note that the lack of shared knowledge between the nginx machines,
coupled with the round robin load balancing between them means the
actual rate limit will be slightly higher, based on the number of
cache machines we have and the load balancing behaviour. Currently,
the worst  case limit is 3 * 9 per minute (1620 per hour).

Note that we have verified the $post_limit_req variable, which is
used as the rate limit, is indeed the client IP address.