Double allowed Sidekiq memory for email-alert-api
Previously we significantly increased the allocated memory for
email-alert-api instances in production [1] in an attempt to
reduce the frequency of worker restarts, which have a tendancy
to lose the work that was being processed (e.g. content changes).
Unfortunately this yielded no improvement, because it's actually
Icinga's thresholds that are triggering the restarts (i.e. not
the allocated memory on the machine). Since the allocated memory
is mostly unused, it should be safe to allow Sidekiq to use more
of it, without impacting other processes. This doubles the max
memory for Sidekiq workers to reduce the frequency of restarts.

[1]: https://github.com/alphagov/govuk-aws/pull/1310