Add a __canary__ response to router
This is intended to simplify the pingdom check. My understanding of the
pingdom canary check is that we want to assert that that GOV.UK origin
is capable of serving responses externally. For example, if we had a
scenario where the data centre hosting origin wasn't on the public
internet Pingdom would alert us.

The check this replaces is quite significantly more complex than this.
It involves a route existing in router-api, Router finding that route
and proxying that request to Nginx on the frontend boxes, and that box
then proxying that request to a backend box which returns a response. I
imagine at one point in time this check made a lot of sense, perhaps it
exhausted much of the stack at that point. However, now it seems strange
that it exercises those particular chunks of the stack. There are alerts
covering these other parts of the stack already so I don't believe this
is necessary for the canary.