Move static canary to assets nginx configuration
This check is used to determine that origin can serve requests on the
assets hostname to the wider internet. This check seemed strangely
placed to be in the static hostname as static is not used publicly via
this hostname (it's only expected to be used by devs and heroku builds).
I felt it made most sense to simplify this to have this check directly
in the assets configuration as that saves any proxying.