Add redirect for Leicestershire lockdown changes
The draft link for this was posted onto twitter by a press organisation
making it rather public. As a temporary measure we are going to redirect
this URL with the token so that anyone following the link ends up on
live GOV.UK.

Once this URL is no-longer in the public domain and we have invalidated
the token we can then remove this redirect with the expectation that
future visitors are taken to a signon login page.