Upgrade to nodejs 12.28
This brings us up to the current LTS version of Node, which will
hopefully buy us some time to sort out trusty so that it can be easier
to install future versions of Node.

From my testing node12 seems to work fine on our trusty machines,
despite lack of official support from nodesource. Because of this it
will be a delicate process deploying this to ensure it doesn't break
statsd.