govuk_solr6: add basic creation of solr "cores"
the `govuk_solr6::core` resource allows us to provision cores. this is
done through solr's http interface, and uses an exec resource to create
the new core, but first checking to see whether a core with this name
already exists. it won't ever try to do anything else as we don't want
to risk puppet screwing around with potentially important data.

include an example core as proof of it working