Add generic webhook plugin to ci master jenkins
This plugin is needed to use trigger application deployments from GitHub
Actions. We are in the process for moving continuous integration to
GitHub Actions. A part of this requires triggering automatic deployments
to successful build of the master branch. This plugin enables Jenkins to
subscribe to incoming webhooks from successful GitHub Action builds and
trigger the "Deploy_App_Downstream" job to deploy the application.