Remove unactionable high load alert
This alert does not signify a definite problem: it's OK for a machine
to be under high load for a period of time, if that happens to be when
we have a lot of work to do. If high load really is a problem, it should
manifest in other, objective alerts. This removes the alert, on the
basis that this data is already available in our monitoring dashboards.