govuk_solr6: add handling of configsets
unlike postgres databases, where each database is permitted to alter its
own schema to its desire, solr (at least in the old fashioned way we're
currentlu forced to use it with ckan) rather wants its schema to be
given to it explicitly "up front". this is typically done by naming a
configSet at core creation time which encompasses a number of settings
for the core, including the schema. as of solr 6.x, these configSets
can't be created through any management interface and have to be placed
in solr's data dir directly. so the `govuk_solr6::configset` resource
allows us to ensure this is in place by simply providing a name and a
`schema.xml`.