Remove unactionable Sidekiq queue size checks
Previously we would alert on the size of the Sidekiq queues, which
was a unique alert for Email Alert API. Since the size of the queue
simply represents the amount of work to do, it cannot be used to
detect a problem - latency is the critical metric here. This removes
all the alerts for the size of Sidekiq queues, which we can still
monitor with our existing Grafana dashboards.