Stop alerting about uninstalled packages
Previously (in 2013) we introduced this alert for outstanding
security updates, on the off-chance they were not automatically
installed by the 'unattended-update', which is run daily [1].
However, the 'apt-check' command we use to detect a lag in updates
also complains about packages that aren't even installed. This
switches the alert to only check for actual updates.

[1]: https://github.com/alphagov/govuk-puppet/commit/105740bb5767c4c4e6490dc7f6a40f9d2b1f27a2