Fix using wrong parameter for delay
Previously we tried to use the 'first_notification_delay' param to
avoid alerting about services that will likely fix themselves. Turns
out this param is only used for the 'notification' feature of Icinga,
not the appearance of the alert in the UI. This replaces the use of
the parameter with the original median averaging technique.

Note that the 'from' param is also reduced, since the calculatio of
the alert state appears to take some kind of average of all the data
in the given range i.e. this is for the alert, not for the viewer.