Only raise 'active keys' warning if keys are active
Currently we raise a warning for all keys older than 90 days.
However, we haven't been checking if keys are actually active;
old keys aren't an issue if they're inactive.

We often inactivate a key while checking if the newly
created key is being used correctly. It would be nice if the
alert didn't continue while we're in that middle state (which can
last for several days).