Remove Email Alert API delivery status checks
These are being removed as these alerts no longer reflect the statuses
within Email Alert API (we have replaced technical_failure and
internal_failure with provider_communication_failure).

We decided that these alerts aren't sufficiently useful to port to the
new status. This is because we regularly pass the thresholds of this
alerts but only momentarily so they very rarely alert despite their low
threshold (1 failure should trigger the alert).