Remove enable_email_alerts setting
We've removed the setting from Travel Advice Publisher so we no longer
need to be setting this environment variable via Puppet. We can't think
of a good reason that we still need this configuration since the
application should behave the same in all environments.

See https://github.com/alphagov/travel-advice-publisher/pull/934 for the
related change in Travel Advice Publisher.