Remove redudant alerts
We decided in [1] that there were some alerts that didn't offer us much
value. These included the checks around a high sidekiq retry set size
and the total email send requests to Notify. Both of these are purely
informational and don't actually indicate that anything is wrong so we
shouldn't be alerting on them.

We also no longer have a daily limit on the amount of emails sent to
Notify with a caveat that Notify itself is currently capped at
delivering 30 million emails a day by amazon SES.

[1] https://docs.google.com/document/d/1fTwyoSD7iWKJYrVk8-yhRSyYBRgZG1jePuW-LzNI74I/edit#heading=h.jvdcecid8r7s

Trello:
https://trello.com/c/TbOhx0B6/438-remove-redundant-alerts