Expand unsafe_machines in all places
Because of the way this code is structured, it modifies the lists before
displaying the results. The `unsafe_machines` array was a copy of the
two `out_of_hours_unsafe_machines` and `in_hours_unsafe_machines` arrays
and therefore modifying it doesn't affectt the original arrays.

This code could do with refactoring but I'll note that down for future
work to get the alert working again.