Fix low rate limit due to broken GitHub Auth
Previously we used a shell variable to set the Authorization header
for the two GitHub requests we make to verify the integrity of a
requested tag, before deploying it downstream. Unfortunately, Jenkins
redacts the API token after the command is run [1], so the variable
literally contained "[******]". This inlines the use of the secret
token in the commands that require it.

[1]: https://www.jenkins.io/blog/2019/02/21/credentials-masking/