Only fail downstream deploy if action required
Previously we marked a build of the Deploy_App_Downstream job as
failed if the goal of the job was not met (to trigger a deploy in
a downstream environment). Since adding Slack notifications, we've
had a number of false positive messages, due to predictable or
unactionable failures. While the goal of the job is not met, the
overall goal of the Continuous Deployment pipeline is:

- We only deploy the latest release of the repo (as we would manually)
- We do not deploy anything if there is a deploy freeze

This switches the exit status for predictable or unactionable failures,
so a build of the job only fails when action is required.