Revert "Revert "Update queue latency check""
The commit applies two changes:

* removes checks for certain Sidekiq queues

This commit removes our icinga checks for email-alert-api's non delivery
related Sidekiq queues. This was decided as part of the work to improve
our alerting system. The remaining delivery queue checks will also be
removed in a future piece of work but can't be yet due to us lacking a
suitable replacement.

* starts checking our delivery transactional queue

We wan't to start checking our delivery_transactional queue as it's used
for one off emails and we should be processing these promptly.