Combine shell commands into a single step
Previously we attempted to use a "0" exit status to silently abort
this job when there are predictable failures. However, having the
exit statements in separate shell steps from the final trigger means
means the trigger (and other steps) are still executed [1]. This
combines the shell steps together, in preparation for changing the
exit behaviour.

[1]: https://github.com/alphagov/govuk-puppet/pull/10600