Cache by host as well as the path ("url")
Previously we only used the path of a request as the cache key,
which leads to surprising outcomes, since the value of the cache
is an absolute URL. This stops overriding the Varnish default [2],
so we cache based on the host of the request, as well as the path.

For example, if a request to 'www-origin/foo' return a redirect
to 'www-origin/bar', we would cache this using '/foo' as the key;
a subsequent reques to 'www/foo' would then return the cached
redirect of 'www-origin/bar', instead of the correct redirect of
'www/bar'. Issues caused by this:

- In production, a cached 'www-origin' redirect would lead to a
network error for users, since 'www-origin' is behind a firewall.

- In integration, a cached 'www-origin' redirect would lead to a
401 Unauthorized response, since it requires basic auth to access.

We think this is likely the cause of many a transient Smokey test
failing in Integration, in steps where we click on a button and
get redirected. Note that Smokey, while does provide the required
basic auth credentials, it only does this for the 'www' domain [1].

[1]: https://varnish-cache.org/docs/3.0/reference/vcl.html
[2]: https://github.com/alphagov/smokey/commit/d8847c4d03b0c416cf9f5d29607184acd566c4fd#diff-57013262a6cdf1f0f6cb2c64aa762c68R49