Remove email_addresses_to_check variable
We no longer want to set this variable here as we can quite easily set
it in the actual application running the check (email-alert-monitoring),
this has the benefit of being much clearer when someone is trying to
understand what the email alert jobs do.

We've also dropped the courtesy copies email as this was deemed
superfluous.

Depends on the PR [here].

[here]: https://github.com/alphagov/email-alert-monitoring/pull/67