Report HTTP errors as critical for healthchecks
Previously we missed a healthcheck alert for Email Alert API due to
the app being down, which manifested as an unknown "timeout" error.
This raises the profile of HTTP errors to critical, since these also
reflect the health of the app being checked.

Conversely, other exceptions are likely to be due to config issues -
not reflective of the app or its environment.