Make AWS LB checks comprehensive
Previously we had a small number of checks for the health of AWS
load balancers, which were incorrectly excluding some hosts, and
had no consistency in choice between internal/external/public

This replaces the current config to represent all of the services
that are load balanced. While we should see other alerts when
individual apps/machines are down, these alerts have the benefit
of also testing the network between the machines. Recently, we
were not alerted in good time when Email Alert API failed to start,
and this should go some way to addressing that, while making the
alerting less arbitrary for all other services.

Where multiple LBs exist for a service, I've used "internal" ones
for consistency, and because the "external" vs. "public" naming
is unclear, and some "-public" LBs appear to be redundant.

Where an Application LB (ALB) is checked, some of the individual
target groups also appear to be redundant. I've excluded these with
a comment to avoid someone thinking this is "correct".

Some LBs are intentionally only associated with a single instance.
In this case, we will only see a critical alert.