Remove redundant check for no processes
This check was introduced to check we were collecting process metrics
for each app [1]. We don't universally check the metrics we collect,
and we have no evidence to suggest this particular metric is a problem.
Note that we have an equivalent check for upstart, which also links
to the same documentation as this alert.

[1]: cd199c36ee642b2b02310259da949ff33ea3a6cc