Show successful/failed Sentry errors over time
#10661 added some useful panels to show the number of Sentry errors
that make it through to Sentry, and the number that is rate-limited,
on a per-environment basis.

However, it did not show this data in the context of time, so it's
hard to know if the figure is improving or worsening. This PR attempts
to remedy that, using the same metrics but now exposing the information
in graphs which show the number of successful/failed requests in
the past 90 days, with a rolling 30 day average so we can see the
general trend.