Fix faulty reload for cache-clearing-service
Previously the command for this service involved running a custom
script that would fail to respond to a SIGHUP, leaving the child
processes running e.g. when Icinga tries to stop it running out of
memory. Since the purpose of the script was just to capture a small
amount of logging output, this removes it from the chain in favour
of a simpler command with no surprising behaviours.