Disable configurations via the Config API in Solr
We have discovered a vulnerability in the version of Solr (6) we
are using for CKAN.

Adding the new `solr.init.sh` startup script allows us to override
the default script (`/etc/default/solr.in.sh`), specifically to disallow
any changes to made to configurations via the Config API. This
also means we cannot use the edit capabilites of the Config API
until further fixes are in place.

According to https://access.redhat.com/security/vulnerabilities/CVE-2017-12629
this should help mitigate the issue, but doesn't remove the XXE
in the first place which we will look into fixing next.