Increase memory thresholds for Asset Manager
This commit increases the memory thresholds for Asset Manager, for both warning and critical. The reason for this bump in the thresholds is after upgrading Asset Manager to Ruby 2.7.1 (from Ruby 2.6.6), which generally results in more memory being consumed. This has been investigated by the Rails upgrade team and seems to be genuine (rather than a memory leak etc), so this commit is bumping the thresholds by 20%, equal to the amount that memory consumption has increased, in order to keep the same (but now increased) baselines.