Stop suppressing non-JSON logs
Previously we replaced any non-JSON logs with an error about how the
non-JSON log could not be decoded. While all logs should ideally be
in the same format, this isn't achievable in practice because we rely
on STDOUT for much of our logging, so are vulnerable to anything doing
an unformatted "puts". This relaxes the error handling by using another
option to ignore non-JSON logs, such that they surface in the "message"
field in Kibana, instead of being lost.

Note that the original behaviour of setting "add_error_key" is retained
when we are not ignoring these errors.