Add a script to monitor unicornherder reloads
We reload Ruby apps, following a deployment, by using upstart to reload
a unicornherder [1] process. This reload creates a new unicorn process
with new workers initialised, after a period of overlap - where the old
and new unicorn processes serve requests - the old process is
terminated.

The problem with this functionality is that we lack visibility of when
a new version of an app is available. We know that we told it to
reload and we expect that some point in the next couple of minutes the
app will be available. This has led to incidents where we've triggered
smoke tests following the deploy of a broken app, but we've been unaware
the app was broken because the smoke tests ran against the old,
non-broken, version of the app.

To counter this problem this commit introduces a wrapper script for
unicornherder reloading. This script monitors whether a new process is
created and when the old one is removed. This was done as a separate
script to unicornherder as there didn't seem a simple synchronous way
to communicate with the process (our existing communication is via signals
which are an asynchronous medium).

The script is expected to be run by a superuser as per the existing reload
commands [2]. To support this I've added in permissions for the deploy user
to be able to run this script with sudo.

There's a couple of parts to the script that we may consider needing to
parameterise should a need arise. It expects the pid file to be in a
consistent location matching the name of the service and it expects the
restarts to complete within 3 minutes.

I wrote this in bash as it seemed to be the most common approach we have
for wrapper scripts and this avoids the dependency version/concerns that
would be the case in Ruby.

Trying this out on a deploy you get the following output:

```
18:03:30   * executing `deploy:restart'
18:03:30   * executing "sudo initctl start email-alert-api 2>/dev/null || sudo govuk_unicorn_reload email-alert-api"
18:03:30     servers: ["ip-10-1-5-9.eu-west-1.compute.internal", "ip-10-1-6-234.eu-west-1.compute.internal"]
18:03:30     [ip-10-1-5-9.eu-west-1.compute.internal] executing command
18:03:30 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Checking email-alert-api is running
18:03:30 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> email-alert-api is running with process 26011. Reloading...
18:03:40 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> New process 26718 started
18:03:40 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Waiting for processes to swap over
18:03:50 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Waiting for processes to swap over
18:04:00 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Waiting for processes to swap over
18:04:10 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Process 26011 has exited
18:04:10 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Successfully reloaded email-alert-api
18:04:10     command finished in 40145ms
18:04:10     [ip-10-1-6-234.eu-west-1.compute.internal] executing command
18:04:10 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Checking email-alert-api is running
18:04:10 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> email-alert-api is running with process 29454. Reloading...
18:04:20 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> New process 30188 started
18:04:20 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Waiting for processes to swap over
18:04:30 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Waiting for processes to swap over
18:04:40 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Waiting for processes to swap over
18:04:50 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Process 29454 has exited
18:04:50 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Successfully reloaded email-alert-api
18:04:50     command finished in 40112ms
```

We can also configure the task to run in parallel to speed up the
process:

```
19:14:16   * executing `deploy:restart'
19:14:16   * executing "sudo initctl start email-alert-api 2>/dev/null || sudo govuk_unicorn_reload email-alert-api"
19:14:16     servers: ["ip-10-1-5-9.eu-west-1.compute.internal", "ip-10-1-6-234.eu-west-1.compute.internal"]
19:14:16     [ip-10-1-6-234.eu-west-1.compute.internal] executing command
19:14:16     [ip-10-1-5-9.eu-west-1.compute.internal] executing command
19:14:16 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Checking email-alert-api is running
19:14:16 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Checking email-alert-api is running
19:14:16 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> email-alert-api is running with process 30188. Reloading...
19:14:16 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> email-alert-api is running with process 26718. Reloading...
19:14:26 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> New process 533 started
19:14:26 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> New process 4064 started
19:14:26 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Waiting for processes to swap over
19:14:26 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Waiting for processes to swap over
19:14:36 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Waiting for processes to swap over
19:14:36 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Waiting for processes to swap over
19:14:46 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Waiting for processes to swap over
19:14:46 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Waiting for processes to swap over
19:14:56 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Process 26718 has exited
19:14:56 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Process 30188 has exited
19:14:56 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] ---> Successfully reloaded email-alert-api
19:14:56 *** [err :: ip-10-1-6-234.eu-west-1.compute.internal] ---> Successfully reloaded email-alert-api
19:14:56     command finished in 40172ms
```

[1]: https://github.com/gds-operations/unicornherder
[2]: https://github.com/alphagov/govuk-app-deployment/blob/96e15410ed5bbd3fe2a5e209dae2d5e55c119ef2/recipes/ruby.rb#L26